Please clang++ by using an explicit cast to unsigned long long.

+++ arangod/V8Server/v8-actions.cpp
@@ -809,8 +809,8 @@ class CallbackTest : public ClusterCommCallback {
     CallbackTest(string msg) : _msg(msg) {}
     virtual ~CallbackTest() {}
     virtual bool operator() (ClusterCommResult* res) {
-      LOG_DEBUG("ClusterCommCallback called on operation %ld",
-                res->operationID);
+      LOG_DEBUG("ClusterCommCallback called on operation %llu",
+                (unsigned long long) res->operationID);
       LOG_DEBUG("Message: %s", _msg.c_str());
       return false;  // Keep it in the queue
     }
